pyproject.toml
[project]
name = "repl-nix-cncordervisualizer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "numpy-stl>=3.1.2",
    "flask-login>=0.6.3",
    "flask-socketio>=5.4.1",
    "cadquery>=2.4.0",
    "ocp>=0.1.4",
    "build123d>=0.7.0",
]
